refactor(subs): directory structure
